fix: resolve component not found error by renaming github button component
