+++ apps/ui-layout/components/website/home/home-header.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import Link from 'next/link';
 import { SearchDialog } from '@/components/website/searchbar';
-import GitHubButton from '@/registry/components/github-repo-btn';
+import GitHubButton from '@/registry/components/github-gradient-button';
 import MobileHeader from '../moibile-header';
 import ThemeSwitch from '../theme-switch';
 import { PlasticButton } from '../ui/plastic-button';
