+++ apps/ui-layout/content/components/github-gradient-button.mdx
@@ -44,7 +44,7 @@ export const metadata = {
   openGraph: {
     title: "Github Gradient Button Component | UI Layouts",
     description: "A GitHub-inspired gradient button component with animated gradient borders, rounded corners, and modern hover effects for stylish button designs.",
-    url: "https://ui-layouts.com/components/uilayouts-github-button",
+    url: "https://ui-layouts.com/components/github-gradient-button",
     images: [
       {
         url: 'https://ui-layouts.com/component-og.jpg',
@@ -78,7 +78,7 @@ export const metadata = {
   category: "technology",
 };
 
-<ComponentCodePreview name='uilayouts-github-button' hasReTrigger isFitheight />
+<ComponentCodePreview name='github-gradient-button' hasReTrigger isFitheight />
 
 > Thanks To [Oni](https://x.com/0ni_x4)
